+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drop support for python < 3.7
 - Drop support for application/msgpackpandas as msgpack support has been dropped from pandas and this is not a standardized content type.
 - Drop support for ujson.
+- Drop support for ini configuration files.
 
 ### Changed
 - Auto update does not request pre-releases by default anymore. It must be explicitly requested via --check_pre_releases.